Updates for Connector framework
