+++ test/SteelToe.Discovery.Client.Test/AbstractBaseTest.cs
@@ -27,7 +27,6 @@ namespace SteelToe.Discovery.Client.Test
             DiscoveryManager.Instance.ClientConfig = null;
             DiscoveryManager.Instance.Client = null;
             DiscoveryManager.Instance.InstanceConfig = null;
-            DiscoveryClientFactory._discoveryClient = null;
         }
     }
 }
